Develop v

+++ src/main/java/com/base/basesetup/dto/CostInvSummaryDTO.java
@@ -1,5 +1,7 @@
 package com.base.basesetup.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -12,14 +14,14 @@ import lombok.NoArgsConstructor;
 public class CostInvSummaryDTO {
 
 	private Long id;
-	private float totChargesBillCurrAmt;
-	private float totChargesLcAmt;
-	private float actBillCurrAmt;
-	private float actBillLcAmt;
-	private float netBillCurrAmt;
-	private float netBillLcAmt;
-	private float roundOff;
-	private float gstInputLcAmt;
+	private BigDecimal totChargesBillCurrAmt;
+	private BigDecimal totChargesLcAmt;
+	private BigDecimal actBillCurrAmt;
+	private BigDecimal actBillLcAmt;
+	private BigDecimal netBillCurrAmt;
+	private BigDecimal netBillLcAmt;
+	private BigDecimal roundOff;
+	private BigDecimal gstInputLcAmt;
